Revert "Reorganize folder structure"
