+++ src/components/EditOverlappingPlayers.tsx
