+++ src/components/GameHistory.tsx
