+++ src/components/InfoModal.tsx
